test_scripts/requirements.txt
alabaster==0.7.13
appdirs==1.4.4
appnope @ file:///opt/concourse/worker/volumes/live/8b5bd665-fe9a-4b43-55ab-604475fcb838/volume/appnope_1606859465017/work
attrs==22.2.0
Babel==2.11.0
backports.zoneinfo==0.2.1
bacpypes==0.17.0
beautifulsoup4==4.12.3
bottle==0.13.4
certifi==2020.6.20
cffi==1.15.1
chacha20poly1305==0.0.3
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.2.1
-e git+https://github.com/aamukta-d/dissertation.git@998ab243cb8c18e5fe07569d4b837f413a39cd30#egg=conpot
copulas==0.7.0
cpppo==5.2.5
crc16==0.1.1
crypto_licensing==5.2.3
cryptography==3.4.8
ctgan==0.5.2
cybox==2.1.0.21
cycler==0.11.0
dataclasses==0.8
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepecho==0.3.0.post1
distlib==0.4.0
dnspython==2.2.1
docutils==0.18.1
e==1.4.5
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
enum34==1.1.10
Faker==14.2.1
filelock==3.4.1
freezegun==1.2.2
fs==2.3.0
future==1.0.0
gevent==22.10.2
graphviz==0.19.1
greenery==2.1
greenlet==2.0.2
hpfeeds3==0.9.10
idna==3.10
imagesize==1.4.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1620913000583/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython==5.8.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
Jinja2==3.0.3
joblib==1.1.1
jupyter-client==5.3.4
jupyter-core==4.5.0
kiwisolver==1.3.1
libtaxii==1.1.119
lxml==5.4.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mixbox==1.0.5
modbus_tk==1.1.5
natsort==8.2.0
numpy==1.19.5
ordered-set==4.0.2
ordereddict==1.1
packaging==21.3
pandas==1.1.5
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
platformdirs==2.4.0
plotly==5.18.0
pluggy==1.0.0
ply==3.11
prompt-toolkit==1.0.15
psutil==5.9.8
ptyprocess @ file:///opt/miniconda3/conda-bld/ptyprocess_1762424170819/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=3470be7f810474c8a2ecfcd6e02acc6aea8483ab595417fa4e336362a349933e
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.21.0
pyghmi==1.4.1
Pygments==2.14.0
pyparsing==3.1.4
pyserial==3.5
pysmi==0.3.2
pysnmp==4.4.12
pysnmp-mibs==0.1.6
pytest==7.0.1
python-dateutil==2.9.0.post0
python-slugify==6.1.2
python-snap7==1.2
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==19.0.2
rdt==1.2.1
requests==2.27.1
requests-toolbelt==1.0.0
scapy==2.4.5
scikit-learn==0.24.2
scipy==1.5.4
sdmetrics==0.8.1
sdv==0.17.2
simplegeneric==0.8.1
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stix==1.2.0.11
stix-validator==3.0.0
tenacity==8.2.2
text-unidecode==1.3
tftpy==0.8.5
threadpoolctl==3.1.0
toml==0.10.2
tomli==1.2.3
torch==1.10.2
torchvision==0.11.3
tornado==5.1.1
tox==3.28.0
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
tzdata==2025.2
tzlocal==4.2
urllib3==1.26.20
virtualenv==20.17.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
weakrefmethod==1.0.3
xlrd==2.0.2
zipp==3.6.0
zope.event==4.6
zope.interface==5.5.2
